Moving component controllers into component file.
